Module type Doubly_linked.S

module Elt : sig ... end

type 'a t

include Ppx_compare_lib.Comparable.S1 with type 'a t := 'a t

val compare : 'a Base__Ppx_compare_lib.compare -> 'a t Base__Ppx_compare_lib.compare

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0__.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0__.Sexp.t) -> 'a t -> Sexplib0__.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

include Base.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc

val fold_result : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'e) Base__.Result.t) -> ('acc, 'e) Base__.Result.t

val fold_until : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'final) Base__Container_intf.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

val exists : 'a t -> f:('a -> bool) -> bool

val for_all : 'a t -> f:('a -> bool) -> bool

val count : 'a t -> f:('a -> bool) -> int

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

val find : 'a t -> f:('a -> bool) -> 'a option

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

include Base.Invariant.S1 with type 'a t := 'a t

val invariant : 'a Base__Invariant_intf.inv -> 'a t Base__Invariant_intf.inv

Creating doubly-linked lists

val create : Base.unit -> 'a t

val of_list : 'a Base.list -> 'a t

of_list l returns a doubly-linked list t with the same elements as l and in the same order (i.e., the first element of l is the first element of t). It is always the case that l = to_list (of_list l).

val of_array : 'a Base.array -> 'a t

Predicates

val equal : 'a t -> 'a t -> Base.bool

pointer equality

val is_first : 'a t -> 'a Elt.t -> Base.bool

val is_last : 'a t -> 'a Elt.t -> Base.bool

val mem_elt : 'a t -> 'a Elt.t -> Base.bool

Constant-time extraction of first and last elements

val first_elt : 'a t -> 'a Elt.t Base.option

val last_elt : 'a t -> 'a Elt.t Base.option

val first : 'a t -> 'a Base.option

val last : 'a t -> 'a Base.option

val first_exn : 'a t -> 'a

val last_exn : 'a t -> 'a

Constant-time retrieval of next or previous element

val next : 'a t -> 'a Elt.t -> 'a Elt.t Base.option

val prev : 'a t -> 'a Elt.t -> 'a Elt.t Base.option

Constant-time insertion of a new element

val insert_before : 'a t -> 'a Elt.t -> 'a -> 'a Elt.t

val insert_after : 'a t -> 'a Elt.t -> 'a -> 'a Elt.t

val insert_first : 'a t -> 'a -> 'a Elt.t

val insert_last : 'a t -> 'a -> 'a Elt.t

Constant-time move of an element from and to positions in the same list

An exception is raised if elt is equal to anchor.

val move_to_front : 'a t -> 'a Elt.t -> Base.unit

val move_to_back : 'a t -> 'a Elt.t -> Base.unit

val move_after : 'a t -> 'a Elt.t -> anchor:'a Elt.t -> Base.unit

val move_before : 'a t -> 'a Elt.t -> anchor:'a Elt.t -> Base.unit

Constant-time removal of an element

val remove : 'a t -> 'a Elt.t -> Base.unit

val remove_first : 'a t -> 'a Base.option

val remove_last : 'a t -> 'a Base.option

val iteri : 'a t -> f:(Base.int -> 'a -> Base.unit) -> Base.unit

val foldi : 'a t -> init:'acc -> f:(Base.int -> 'acc -> 'a -> 'acc) -> 'acc

val fold_elt : 'a t -> init:'acc -> f:('acc -> 'a Elt.t -> 'acc) -> 'acc

fold_elt t ~init ~f is the same as fold, except f is called with the 'a Elt.t's from the list instead of the contained 'a values.

Note that like other iteration functions, it is an error to mutate t inside the fold. If you'd like to call remove on any of the 'a Elt.t's, use filter_inplace.

val foldi_elt : 'a t -> init:'acc -> f:(Base.int -> 'acc -> 'a Elt.t -> 'acc) -> 'acc

val iter_elt : 'a t -> f:('a Elt.t -> Base.unit) -> Base.unit

val iteri_elt : 'a t -> f:(Base.int -> 'a Elt.t -> Base.unit) -> Base.unit

val fold_right : 'a t -> init:'acc -> f:('a -> 'acc -> 'acc) -> 'acc

val fold_right_elt : 'a t -> init:'acc -> f:('a Elt.t -> 'acc -> 'acc) -> 'acc

val find_elt : 'a t -> f:('a -> Base.bool) -> 'a Elt.t Base.option

find_elt t ~f finds the first element in t that satisfies f, by testing each of element of t in turn until f succeeds.

val findi_elt : 'a t -> f:(Base.int -> 'a -> Base.bool) -> (Base.int * 'a Elt.t) Base.option

val clear : 'a t -> Base.unit

clear t removes all elements from the list in constant time.

val copy : 'a t -> 'a t

val transfer : src:'a t -> dst:'a t -> Base.unit

transfer ~src ~dst has the same behavior as iter src ~f:(insert_last dst); clear src except that it runs in constant time.

If s = to_list src and d = to_list dst, then after transfer ~src ~dst:

to_list src = []

to_list dst = d @ s

Linear-time mapping of lists (creates a new list)

val map : 'a t -> f:('a -> 'b) -> 'b t

val mapi : 'a t -> f:(Base.int -> 'a -> 'b) -> 'b t

val filter : 'a t -> f:('a -> Base.bool) -> 'a t

val filteri : 'a t -> f:(Base.int -> 'a -> Base.bool) -> 'a t

val filter_map : 'a t -> f:('a -> 'b Base.option) -> 'b t

val filter_mapi : 'a t -> f:(Base.int -> 'a -> 'b Base.option) -> 'b t

Linear-time partition of lists (creates two new lists)

val partition_tf : 'a t -> f:('a -> Base.bool) -> 'a t * 'a t

val partitioni_tf : 'a t -> f:(Base.int -> 'a -> Base.bool) -> 'a t * 'a t

val partition_map : 'a t -> f:('a -> ('b, 'c) Base.Either.t) -> 'b t * 'c t

val partition_mapi : 'a t -> f:(Base.int -> 'a -> ('b, 'c) Base.Either.t) -> 'b t * 'c t

Linear-time in-place mapping of lists

val map_inplace : 'a t -> f:('a -> 'a) -> Base.unit

map_inplace t ~f replaces all values v with f v

val mapi_inplace : 'a t -> f:(Base.int -> 'a -> 'a) -> Base.unit

val filter_inplace : 'a t -> f:('a -> Base.bool) -> Base.unit

filter_inplace t ~f removes all elements of t that don't satisfy f.

val filteri_inplace : 'a t -> f:(Base.int -> 'a -> Base.bool) -> Base.unit

val filter_map_inplace : 'a t -> f:('a -> 'a Base.option) -> Base.unit

If f returns None, the element is removed, else the value is replaced with the contents of the Some

val filter_mapi_inplace : 'a t -> f:(Base.int -> 'a -> 'a Base.option) -> Base.unit

val unchecked_iter : 'a t -> f:('a -> Base.unit) -> Base.unit

unchecked_iter t ~f behaves like iter t ~f except that f is allowed to modify t. Adding or removing elements before the element currently being visited has no effect on the traversal. Elements added after the element currently being visited will be traversed. Elements deleted after the element currently being visited will not be traversed. Deleting the element currently being visited is an error that is not detected (presumably leading to an infinite loop).

val to_sequence : 'a t -> 'a Base.Sequence.t

A sequence of values from the doubly-linked list. It makes an intermediate copy of the list so that the returned sequence is immune to any subsequent mutation of the original list.